moved create_package_list to dpkg_integration.rs also added serde rename_all on struct Package
